Changing S5 models.py file to a directory
This addresses Issue #90 but is as of yet untested
