+++ eureka/S5_lightcurve_fitting/models/__init__.py
@@ -1,0 +1,5 @@
+from . import Model
+from .CompositeModel import CompositeModel
+from .TransitModel import TransitModel
+from .PolynomialModel import PolynomialModel
+from .ExpRampModel import ExpRampModel
